MinDownPayment working
Finished setting up MinDownPayment API.
Added css to NewPayment and PaymentForm.

TODO:
Separation of responsibilities: PaymentForm component is currently too big. It is responsible for both sending and receiving data from the API call.
PaymentForm should ONLY be responsible for the payment form, it does not need to know about min down payment.
Need to make some components reusable so that it can be reuse for Amortization feature.